+++ api/sql/schema/340-rule.conditionActorCustom.sql
@@ -1,0 +1,5 @@
+CREATE TYPE [rule].[conditionActorCustom] AS TABLE(
+    conditionName NVARCHAR(100),
+    factor CHAR(2),
+    actorId BIGINT
+)