+++ api/sql/schema/340-rule.properties.sql
@@ -1,5 +1,5 @@
 CREATE TYPE [rule].[properties] AS TABLE (
     [factor] CHAR(2),
-    [name] NVARCHAR (100),
-    [value] BIGINT
+    [name] NVARCHAR(50),
+    [value] NVARCHAR(200)
 )