+++ api/sql/schema/750-rule.rule.add.sql
@@ -12,16 +12,22 @@ DECLARE @splitName [rule].splitNameTT,
     @splitAssignment [rule].splitAssignmentTT,
     @conditionId INT
 BEGIN TRY
+    -- checks if the user has a right to make the operation
+    DECLARE @actionID VARCHAR(100) = OBJECT_SCHEMA_NAME(@@PROCID) + '.' + OBJECT_NAME(@@PROCID), @return INT = 0
+    EXEC @return = [user].[permission.check] @actionId = @actionID, @objectId = NULL, @meta = @meta
+    IF @return != 0
+    BEGIN
+        RETURN 55555
+    END
 
     IF EXISTS
         (
-            SELECT [priority]
+            SELECT [name]
             FROM [rule].condition
-            WHERE [priority] = (SELECT [priority] FROM @condition)
-                AND isDeleted = 0
+            WHERE [name] = (SELECT [name] FROM @condition)
         )
         BEGIN
-            RAISERROR ('rule.duplicatedPriority', 16, 1)
+            RAISERROR ('rule.duplicatedName', 16, 1)
         END
 
     BEGIN TRANSACTION
@@ -32,6 +38,9 @@ BEGIN TRY
             operationEndDate,
             sourceAccountId,
             destinationAccountId,
+            [name],
+            [description],
+            notes,
             createdOn,
             createdBy
             )
@@ -41,7 +50,10 @@ BEGIN TRY
             operationEndDate,
             sourceAccountId,
             destinationAccountId,
-            GETDATE(),
+            [name],
+            [description],
+            notes,
+            GETUTCDATE(),
             ISNULL(createdBy, @userId)
         FROM @condition;
 
@@ -228,7 +240,7 @@ BEGIN TRY
         SELECT
             @conditionId [key],
             c.priority rulePriority,
-            GETDATE() creationDateTime
+            GETUTCDATE() creationDateTime
         FROM
             @condition c
         FOR XML RAW