+++ api/sql/schema/750-rule.rule.remove.sql
@@ -6,67 +6,9 @@ DECLARE @userId BIGINT = (SELECT [auth.actorId] FROM @meta)
 BEGIN TRY
 
     BEGIN TRANSACTION
-        DELETE x
-        FROM
-            [rule].limit x
-        JOIN
-            @conditionId item ON x.conditionId = item.value
-
-        DELETE x
-        FROM
-            [rule].conditionActor x
-        JOIN
-            @conditionId item ON x.conditionId = item.value
-
-        DELETE x
-        FROM
-            [rule].conditionItem x
-        JOIN
-            @conditionId item ON x.conditionId = item.value
-
-        DELETE x
-        FROM
-            [rule].conditionProperty x
-        JOIN
-            @conditionId item ON x.conditionId = item.value
-
-        DELETE
-            x
-        FROM
-            [rule].splitRange x
-        JOIN
-            [rule].splitName s ON s.splitNameId = x.splitNameId
-        JOIN
-            @conditionId item ON s.conditionId = item.value
-        DELETE
-            x
-        FROM
-            [rule].splitAnalytic x
-        JOIN
-            [rule].splitAssignment y ON y.splitAssignmentId = x.splitAssignmentId
-        JOIN
-            [rule].splitName s ON s.splitNameId = y.splitNameId
-        JOIN
-            @conditionId item ON s.conditionId = item.value
-
-        DELETE
-            x
-        FROM
-            [rule].splitAssignment x
-        JOIN
-            [rule].splitName s ON s.splitNameId = x.splitNameId
-        JOIN
-            @conditionId item ON s.conditionId = item.value
-        DELETE
-            x
-        FROM
-            [rule].splitName x
-        JOIN
-            @conditionId item ON x.conditionId = item.value
-
         UPDATE x
         SET isDeleted = 1,
-        updatedOn = GETDATE(),
+        updatedOn = GETUTCDATE(),
         updatedBy = @userId
         FROM
             [rule].condition x
@@ -77,7 +19,7 @@ BEGIN TRY
         SELECT
             x.conditionId [key],
             x.priority rulePriority,
-            GETDATE() deletionDateTime
+            GETUTCDATE() deletionDateTime
         FROM
             [rule].condition x
         JOIN