+++ errors.js
@@ -4,6 +4,7 @@ module.exports = function error({utError: {defineError, getError, fetchErrors}})
         const Rule = defineError('rule', null, 'ut-rule rule error', 'error');
 
         defineError('generic', Rule, 'Rule generic');
+        defineError('amount', Rule, 'Invalid amount');
         defineError('ruleNotExists', Rule, 'Rule does not exists');
         defineError('exceedMaxLimitAmount', Rule, 'Transaction amount is above maximum');
         defineError('exceedMinLimitAmount', Rule, 'Transaction amount is below minimum');
@@ -17,6 +18,7 @@ module.exports = function error({utError: {defineError, getError, fetchErrors}})
         defineError('exceedWeeklyLimitCount', Rule, 'Weekly count limit reached');
         defineError('exceedMonthlyLimitCount', Rule, 'Monthly count limit reached');
         defineError('duplicatedPriority', Rule, 'Rule with this priority already exists');
+        defineError('duplicatedName', Rule, 'Rule with this name already exists');
         defineError('securityViolation', Rule, 'Unauthorized operation');
         defineError('unauthorizedMinLimitAmount', Rule, 'Transaction amount below minimum unauthorized');
         defineError('unauthorizedMaxLimitAmount', Rule, 'Transaction amount above maximum unauthorized');