+++ ui/react/components/Grid/index.js
@@ -116,7 +116,7 @@ export default class Grid extends React.Component {
 
     updateColumns = (columns) => {
         this.setState({
-            columns: columns
+            columns
         });
     };
 
@@ -192,8 +192,20 @@ export default class Grid extends React.Component {
                 case 'updatedOn':
                     return this.renderGridColumn(value, ['updatedOn'], row, 'updatedOn');
                 case 'expansion': {
-                    const expansionText = this.state.expandedGridColumns.some(v => v === row.priority) ? 'See less...' : 'See more...';
-                    return <a onClick={(e) => { e.preventDefault(); this.handleGridExpansion(row.priority); }}>{expansionText}</a>;
+                    let showText;
+                    Object.values(row?.channel).every(data => {
+                        if (data.length >= 10) {
+                            showText = true;
+                            return false;
+                        }
+                        return true;
+                    });
+                    if (showText) {
+                        const expansionText = this.state.expandedGridColumns.some(v => v === row.priority) ? 'See less...' : 'See more...';
+                        return <a onClick={(e) => { e.preventDefault(); this.handleGridExpansion(row.priority); }}>{expansionText}</a>;
+                    } else {
+                        return '';
+                    }
                 }
                 default:
                     return value;