+++ validations/rule/rule.condition.edit.js
@@ -1,0 +1,10 @@
+// @ts-check
+/** @type { import("ut-run").validationFactory } */
+module.exports = ({
+    joi
+}) => ({
+    'rule.condition.edit': () => ({
+        params: joi.any(),
+        result: joi.any()
+    })
+});